requirements.txt
gitpython
git+https://github.com/ahharu/stashy.git
jenkinsapi